openvino_serving/requirements.txt
numpy
Flask
opencv-contrib-python-headless
pillow
openvino-dev
ovmsclient